refactor(memory): split the reply pipeline's three subsystems out of one file
Closes #607
