+++ src/discordbot/services/memory/pipeline.py
@@ -1,35 +1,45 @@
-"""Background orchestration for the two-phase memory pipeline.
+"""Per-scope orchestration for the memory pipeline.
 
 The pipeline is keyed by an opaque scope (see ``store``), so the same
 orchestration drives both per-user and per-server (bot self) memory. The
 flavor-specific bits are injected: ``subject`` names the memory target and
 ``writer`` carries the flavor's prompts.
+
+What lives here is the part that binds the rest together: one turn's review, the
+consolidation fan-out that turn can trigger, and the clear that has to neutralise every
+tier at once. The three subsystems that touch almost nothing else moved out beside it in
+#607 — ``inflight`` holds the one-job-per-scope queue and the reply.db bookkeeping,
+``regeneration`` the from-scratch rebuild, ``tone`` the unpartitioned tone tier.
+
+They do not all sit on the same side of this module. ``inflight`` and ``tone`` are below it
+and import nothing from here — the turn body ``inflight`` runs reaches it as an argument
+rather than an import, which is what keeps that edge one-way. ``regeneration`` is ABOVE it
+and takes the compartment machinery (``CompartmentInput``, ``compartment_request``,
+``apply_forget_buckets``, ``global_first``, ``memory_semaphore``, ``report_injection_size``)
+from here, so nothing in this module may import it back.
 """
 
 import time
-from typing import Literal
 import asyncio
 from datetime import UTC, datetime
-from collections.abc import Callable, Awaitable
 
 import logfire
-from pydantic import Field, BaseModel, ConfigDict, SkipValidation
+from pydantic import Field, BaseModel, ConfigDict
 from openai.types.responses.response_input_param import EasyInputMessageParam
 
 from discordbot.typings.memory import MemoryOwner, MemoryWriteSummary
 from discordbot.services.memory import database as memory_db
 from discordbot.typings.timeouts import MEMORY_CONSOLIDATE_TIMEOUT_SECONDS
-from discordbot.utils.asyncio_locks import KeyedLockManager, LoopLocalRegistry, LoopLocalSemaphore
+from discordbot.utils.asyncio_locks import LoopLocalSemaphore
+from discordbot.services.memory.tone import update_tone_note
 from discordbot.services.memory.facts import MemoryFlavor, parse_identity, sections_for_flavor
 from discordbot.services.memory.store import (
     DM_COMPARTMENT,
     GLOBAL_COMPARTMENT,
     clear_raw,
-    read_tone,
-    clear_tone,
+    flavor_of,
     read_facts,
     scope_lock,
-    write_tone,
     mark_cleared,
     append_detail,
     cleared_since,
@@ -39,9 +49,7 @@ from discordbot.services.memory.store import (
     read_detail_tail,
     read_raw_entries,
     count_raw_entries,
-    detail_file_bytes,
     list_compartments,
-    prune_compartment,
     delete_memory_files,
     read_memory_document,
 )
@@ -52,13 +60,11 @@ from discordbot.services.memory.deltas import (
     sweep_stale_facts,
     partition_raw_entries,
     render_existing_facts,
-    tone_evidence_from_raw,
     partition_forget_requests,
 )
 from discordbot.services.memory.writer import (
     MemoryWriterAI,
     MemoryObservation,
-    ConsolidatedMemory,
     ConsolidationRequest,
     parse_turn_payload,
     render_turn_payload,
@@ -69,117 +75,32 @@ from discordbot.services.memory.writer import (
     filter_duplicate_observations,
 )
 from discordbot.typings.context_budgets import (
-    MEMORY_MERGED_NOTES_MAX,
     MEMORY_INJECTION_MAX_CHARS,
     MEMORY_INJECTION_WARN_CHARS,
     MEMORY_DETAIL_CONTEXT_MAX_CHARS,
+)
+from discordbot.services.memory.inflight import (
+    MemoryTurn,
+    MemoryWriteReport,
+    stage_turn,
+    report_writes,
+    safe_db_write,
+    staging_locks,
+    drop_pending_updates,
+    enqueue_memory_update,
 )
 from discordbot.services.memory.constants import (
     COMPACTION_TRIGGER_CHARS,
     MEMORY_GLOBAL_CONCURRENCY,
     RAW_CONSOLIDATION_MAX_BYTES,
     RAW_CONSOLIDATION_THRESHOLD,
-    MEMORY_REGENERATION_COOLDOWN_SECONDS,
     MEMORY_CONSOLIDATION_COOLDOWN_SECONDS,
 )
 from discordbot.services.memory.git_history import memory_git
 
-# What a caller is handed once a turn's memory writes land. `services/` never composes what
-# a user reads, so this reports the shape and lets the cog word it. In-memory only: a resumed
-# turn runs after a restart, long past the reply it belonged to, and has no report to make.
-type MemoryWriteReport = Callable[[MemoryWriteSummary], Awaitable[None]]
-
-# The ways a from-scratch rebuild can end, carried on `RegenerationReport.result`.
-_RegenerationResult = Literal["regenerated", "no_evidence", "failed", "cooldown"]
-
-
-class RegenerationReport(BaseModel):
-    """What one from-scratch rebuild did, for a caller with no logfire to read.
-
-    Attributes:
-        result: How the rebuild ended.
-        unreadable_removed: Fact files it destroyed that no reader could parse.
-    """
-
-    model_config = ConfigDict(frozen=True)
-
-    result: _RegenerationResult = Field(..., description="How the rebuild ended.")
-    unreadable_removed: int = Field(
-        default=0, description="Fact files removed that no reader could parse."
-    )
-
-
-class _PendingMemoryUpdate(BaseModel):
-    """The newest skipped update request, replayed once the in-flight task ends.
-
-    Attributes:
-        subject: The phase-1 directive naming the memory target.
-        transcript: The rendered phase-1 input captured for the skipped turn
-            (already folds in the reply), so the replay needs no re-render.
-        writer: The memory writing service to run the replayed update with.
-        identity: Single-line target identity `parse_identity` splits into the
-            `owner_id` / `owner_name` stamped on every fact this scope writes.
-        captured_at: `time.monotonic()` when the turn was captured, so a clear
-            that lands before the replay can abort it via `cleared_since`.
-        token: Process-local logical token persisted with the deferred turn's DB
-            row, reused on replay so the terminal write guards on the same id.
-    """
-
-    model_config = ConfigDict(arbitrary_types_allowed=True)
-
-    subject: str = Field(..., description="The phase-1 directive naming the memory target.")
-    transcript: str = Field(
-        ..., description="The rendered phase-1 input captured for the skipped turn."
-    )
-    writer: SkipValidation[MemoryWriterAI] = Field(
-        ..., description="The memory writing service to run the replayed update with."
-    )
-    identity: str = Field(
-        ...,
-        description=(
-            "Single-line target identity `parse_identity` splits into the `owner_id` / "
-            "`owner_name` stamped on every fact this scope writes."
-        ),
-    )
-    captured_at: float = Field(
-        ...,
-        description=(
-            "`time.monotonic()` when the turn was captured, so a clear that lands "
-            "before the replay can abort it via `cleared_since`."
-        ),
-    )
-    token: int = Field(
-        ..., description="Logical version token reused on replay for the DB row guard."
-    )
-    report: SkipValidation[MemoryWriteReport | None] = Field(
-        default=None, description="Callback reporting what the replayed turn recorded."
-    )
-
-
-# Process-level per-scope in-flight de-dupe; while one update runs, the skipped turns are
-# held and replayed afterwards. Within ONE conversation source only the newest is kept, its
-# history window already covering the earlier ones, and its memory notes merged in so nothing
-# a marker wrote is lost.
-#
-# The second key is the subject, which carries the source line, and it is a correctness
-# boundary rather than bookkeeping: the scope is guild-independent, so a user active in two
-# guilds at once would otherwise have one conversation's notes replayed under the other's
-# source stamp, filing a `source_only` observation in a compartment the speaker never spoke
-# in. Sources are replayed one after another, each keeping its own subject.
-#
-# Loop-local, like `_regeneration_tasks` below: an `asyncio.Task` belongs to the loop that
-# created it, so an entry surviving a loop change parks every later turn for that scope
-# behind a task nothing on this loop can ever see finish.
-_inflight_tasks: LoopLocalRegistry[str, asyncio.Task[None]] = LoopLocalRegistry()
-_pending_updates: LoopLocalRegistry[str, dict[str, _PendingMemoryUpdate]] = LoopLocalRegistry()
-
 # Per-scope consolidation attempt times for the cooldown; monotonic, so it does
 # not need a loop-change reset. Tests clear it through the conftest fixture.
 _last_consolidation: dict[str, float] = {}
-
-# The exact header a tone note must lead with; the tier is injected on every reply,
-# so anything else is a rewrite that did not land and must not be written.
-_TONE_HEADER = "## 語氣偏好"
 
 # Consecutive refused consolidation batches per (scope, compartment). A refusal keeps the
 # raw batch for retry, which is right for a transient LLM failure and wrong for a
@@ -189,90 +110,20 @@ _TONE_HEADER = "## 語氣偏好"
 _consecutive_rejections: dict[tuple[str, str], int] = {}
 _MAX_QUIET_REJECTIONS = 3
 
-# Per-scope regeneration attempt times, separate from the consolidation cooldown
-# so a manual `/memory regenerate` never starves the automatic background
-# consolidation or vice versa. Recorded at attempt time so failures cool down too.
-_last_regeneration: dict[str, float] = {}
-
-# Per-scope in-flight regeneration tasks so a manual rebuild runs in the
-# background without blocking the command, and a second request while one is
-# still running cannot double-schedule the rebuild. Kept separate
-# from `_inflight_tasks` because regeneration is a distinct, user-triggered job.
-_regeneration_tasks: LoopLocalRegistry[str, asyncio.Task[RegenerationReport]] = LoopLocalRegistry()
-
 # Process-wide semaphore capping concurrent background memory updates so a busy server
 # cannot fan out unbounded LLM work; shared across flavors and rebuilt per loop. The cap
 # is read at build time so a test that lowers MEMORY_GLOBAL_CONCURRENCY first still applies.
 _memory_semaphore_holder = LoopLocalSemaphore(capacity_provider=lambda: MEMORY_GLOBAL_CONCURRENCY)
 
 
-def _memory_semaphore() -> asyncio.Semaphore:
+def memory_semaphore() -> asyncio.Semaphore:
     """Returns the process-wide semaphore, rebuilt when the event loop changes."""
     return _memory_semaphore_holder.get()
 
 
-# Detached best-effort reply.db writes (the deferred-turn persist), held so the
-# event loop keeps a strong reference until they finish; reset by the test fixture.
-_db_tasks: set[asyncio.Task[None]] = set()
-
-# A clear and the short reply.db staging transaction must not pass each other:
-# otherwise an INSERT can commit after the clear's tombstone write. This is separate
-# from the minutes-long file-write lock and is held only around reply.db staging
-# or the clear's tombstone plus synchronous file removal.
-_staging_locks = KeyedLockManager[str]()
-
-
-def flavor_of(scope: str) -> memory_db.MemoryJobFlavor:
-    """Maps a scope to its persisted memory flavor (`server_scope` carries a '/')."""
-    return "server" if "/" in scope else "user"
-
-
-async def _safe(coro: Awaitable[None]) -> None:
-    """Awaits a best-effort reply.db write, swallowing any failure.
-
-    Persistence is an augmentation layer (the opposite of `cogs/research/cog.py`,
-    which lets DB errors raise into its run loop): a reply.db failure must never
-    break the in-memory fire-and-forget memory pipeline.
-    """
-    try:
-        await coro
-    except Exception:
-        logfire.warn("memory_job persistence write failed", _exc_info=True)
-
-
-def _spawn_db(coro: Awaitable[None]) -> None:
-    """Runs a detached best-effort DB write, tracked so it is not GC'd mid-flight."""
-    task = asyncio.ensure_future(_safe(coro=coro))
-    _db_tasks.add(task)
-    task.add_done_callback(_db_tasks.discard)
-
-
-async def _stage_turn(  # noqa: PLR0913 -- one row's columns plus the turn's capture time
-    *, scope: str, subject: str, transcript: str, identity: str, token: int, captured_at: float
-) -> None:
-    """Stages one turn only in the memory lifetime that captured it.
-
-    The short per-scope lock serializes staging with the clear's tombstone write.
-    A row committed before a clear is scrubbed by that newer logical token. A row
-    waiting while a clear is in progress sees its closing stamp before it can
-    write, so it cannot leave a during-clear transcript for restart to resume.
-    """
-    async with _staging_locks.hold(key=scope):
-        if cleared_since(scope=scope, started_at=captured_at):
-            return
-        await memory_db.upsert_pending(
-            scope=scope,
-            flavor=flavor_of(scope=scope),
-            subject=subject,
-            transcript=transcript,
-            identity=identity,
-            token=token,
-        )
-
-
 async def _clear_scope_critical(scope: str) -> tuple[bool, bool]:
     """Completes the non-interruptible durable portion of one memory clear."""
-    async with _staging_locks.hold(key=scope):
+    async with staging_locks.hold(key=scope):
         removed_job = await memory_db.clear_job(
             scope=scope, flavor=flavor_of(scope=scope), token=memory_db.new_token()
         )
@@ -350,12 +201,11 @@ async def clear_scope_memory(scope: str) -> bool:
     # clear on its own re-check. Ordering the stamp first is what makes the
     # tombstone below safe without draining or taking the minutes-long scope lock.
     mark_cleared(scope=scope)
-    # Drops the retained transcript now rather than waiting for the in-flight
-    # task to finish and discard it; `_finish_memory_update` then finds no
-    # pending turn and replays nothing. Each dropped turn's reply is told, or it
-    # would keep saying it was still working on memory this clear just erased.
-    for dropped in (_pending_updates.pop(key=scope) or {}).values():
-        _release_pending_report(pending=dropped)
+    # Drops the retained transcripts now rather than waiting for the in-flight
+    # task to finish and discard them; `inflight` then finds no pending turn and
+    # replays nothing. Each dropped turn's reply is told, or it would keep saying
+    # it was still working on memory this clear just erased.
+    drop_pending_updates(scope=scope)
     critical_task = asyncio.create_task(_clear_scope_critical(scope=scope))
     (removed_files, removed_job), caller_cancelled = await _await_clear_critical(
         task=critical_task
@@ -418,16 +268,19 @@ def schedule_memory_update(  # noqa: PLR0913 -- flavor (scope/subject/identity) 
     if not remember_notes and not forget_notes:
         return
     transcript = transcript_from_messages(message_list=message_list, full_reply=full_reply)
-    _enqueue_memory_update(
-        scope=scope,
-        subject=subject,
-        transcript=render_turn_payload(
-            transcript=transcript, remember=remember_notes, forget=forget_notes
+    enqueue_memory_update(
+        turn=MemoryTurn(
+            scope=scope,
+            subject=subject,
+            transcript=render_turn_payload(
+                transcript=transcript, remember=remember_notes, forget=forget_notes
+            ),
+            writer=writer,
+            identity=identity,
+            token=memory_db.new_token(),
+            report=report,
         ),
-        writer=writer,
-        identity=identity,
-        token=memory_db.new_token(),
-        report=report,
+        run=_run_memory_update,
     )
 
 
@@ -435,75 +288,17 @@ def resume_memory_update(  # noqa: PLR0913 -- mirrors a persisted row's columns
     *, scope: str, subject: str, transcript: str, writer: MemoryWriterAI, identity: str, token: int
 ) -> None:
     """Re-enqueues a persisted phase-1 turn on restart, reusing its stored token."""
-    _enqueue_memory_update(
-        scope=scope,
-        subject=subject,
-        transcript=transcript,
-        writer=writer,
-        identity=identity,
-        token=token,
-    )
-
-
-def _enqueue_memory_update(  # noqa: PLR0913 -- flavor (scope/subject/identity) plus the rendered turn
-    scope: str,
-    subject: str,
-    transcript: str,
-    writer: MemoryWriterAI,
-    identity: str,
-    token: int,
-    report: MemoryWriteReport | None = None,
-) -> None:
-    """Schedules (or defers) one rendered-transcript update, backed by a reply.db row."""
-    # Stamped here, not inside the worker: a clear landing between this call and
-    # the task actually starting must still abort the turn, and a worker that
-    # timed itself would read the clear as older than its own work and write on.
-    captured_at = time.monotonic()
-    running = _inflight_tasks.get(key=scope)
-    if running is not None and not running.done():
-        by_subject = _pending_updates.setdefault(key=scope, default={})
-        superseded = by_subject.get(subject)
-        if superseded is not None:
-            transcript = _merged_payload(newer=transcript, older=superseded.transcript)
-            report = _merged_report(newer=report, older=superseded.report)
-        by_subject[subject] = _PendingMemoryUpdate(
-            subject=subject,
-            transcript=transcript,
-            report=report,
-            writer=writer,
-            identity=identity,
-            captured_at=captured_at,
-            token=token,
-        )
-        # Persist the deferred turn so a redeploy before it runs still resumes it.
-        # Safe from a same-token race: this turn's worker only starts after the
-        # in-flight one ends, long after this detached write lands, and it carries
-        # a newer token than the running turn so newest-wins keeps it.
-        _spawn_db(
-            coro=_stage_turn(
-                scope=scope,
-                subject=subject,
-                transcript=transcript,
-                identity=identity,
-                token=token,
-                captured_at=captured_at,
-            )
-        )
-        return
-    task = asyncio.create_task(
-        _run_memory_update(
+    enqueue_memory_update(
+        turn=MemoryTurn(
             scope=scope,
             subject=subject,
             transcript=transcript,
             writer=writer,
             identity=identity,
             token=token,
-            captured_at=captured_at,
-            report=report,
-        )
+        ),
+        run=_run_memory_update,
     )
-    _inflight_tasks.set(key=scope, value=task)
-    task.add_done_callback(lambda finished: _finish_memory_update(scope=scope, task=finished))
 
 
 def _write_summary(
@@ -527,155 +322,7 @@ def _write_summary(
     )
 
 
-async def _report_writes(report: MemoryWriteReport, summary: MemoryWriteSummary) -> None:
-    """Hands the caller what this turn recorded, never letting the report cost the write.
-
-    The write is already durable by the time this runs, so a Discord edit that fails, or a
-    reply that has since been deleted, must not surface as a memory failure.
-
-    An empty summary is reported like any other. It is the answer to a real question — the
-    turn is over and kept nothing — and the caller is holding a reply that says the work is
-    still going, so silence here reads as the work never having finished.
-    """
-    try:
-        await report(summary)
-    except Exception:
-        # Broad on purpose: the callback reaches Discord, and this runs in a background task
-        # whose failure would otherwise be logged as the memory update having failed.
-        logfire.warn("Reporting a memory write back to the reply failed", _exc_info=True)
-
-
-def _merged_payload(newer: str, older: str) -> str:
-    """Carries a superseded turn's memory notes into the turn replacing it.
-
-    Only the newest skipped turn is replayed, and for a TRANSCRIPT that is right: its history
-    window already contains the earlier skipped turns, which is what makes one replay enough. A
-    marker note is not in that window. It exists only in the reply that emitted it, so letting
-    the newer payload simply overwrite the older one would drop it with nothing in the logs to
-    say a note had ever been written.
-
-    Merging is only ever within one conversation source; `_pending_updates` is keyed on the
-    subject for that reason, and this function never sees two sources.
-    """
-    transcript, remember, forget = parse_turn_payload(payload=newer)
-    _, older_remember, older_forget = parse_turn_payload(payload=older)
-    return render_turn_payload(
-        transcript=transcript,
-        remember=_deduped(notes=(*older_remember, *remember)),
-        forget=_deduped(notes=(*older_forget, *forget)),
-    )
-
-
-def _merged_report(
-    newer: MemoryWriteReport | None, older: MemoryWriteReport | None
-) -> MemoryWriteReport | None:
-    """Answers both replies when one turn's notes are merged into another's.
-
-    The counterpart to `_merged_payload`, and needed for the same reason: the superseded turn's
-    notes ride on into the merged payload, so its reply is still waiting to be told what became
-    of them. Overwriting the report the way the transcript is overwritten would leave that reply
-    saying `正在整理記憶⋯` for good.
-
-    Both are handed the same summary, which is the only honest one available: the two turns'
-    notes were reviewed as one payload and there is nothing left that says which note came from
-    which reply.
-
-    Each is answered independently. A merge exists because the older reply is still standing
-    there promising work, and that reply is also the likelier of the two to have been deleted
-    under it, so a callback that raises must not take the other one's report with it.
-    """
-    if newer is None or older is None:
-        return newer or older
-
-    async def both(summary: MemoryWriteSummary) -> None:
-        """Reports one merged outcome to every reply whose notes went into it."""
-        await _report_writes(report=older, summary=summary)
-        await _report_writes(report=newer, summary=summary)
-
-    return both
-
-
-def _deduped(notes: tuple[str, ...]) -> tuple[str, ...]:
-    """Drops exact repeats, keeps writing order, and caps what a merge can accumulate.
-
-    The per-reply cap the markers enforce does not survive a merge: each skipped turn stacks
-    onto the pending payload, so without a bound here a long stretch of skipped turns grows
-    the review request and the raw entry together. The OLDEST are dropped, since the newest
-    notes are the ones the user is still in the middle of.
-    """
-    unique = tuple(dict.fromkeys(notes))
-    return unique[-MEMORY_MERGED_NOTES_MAX:]
-
-
-def _release_pending_report(pending: _PendingMemoryUpdate) -> None:
-    """Tells a dropped turn's reply that nothing was recorded, detached.
-
-    A deferred turn is dropped rather than replayed when the scope was cleared under it, and
-    its reply is still showing `正在整理記憶⋯`. Detached because both callers are synchronous —
-    a done-callback and the clear orchestration — while the report reaches Discord.
-    """
-    if pending.report is None:
-        return
-    _spawn_db(coro=_report_writes(report=pending.report, summary=MemoryWriteSummary()))
-
-
-def _finish_memory_update(scope: str, task: asyncio.Task[None]) -> None:
-    """Clears the in-flight slot, logs failures, and replays a pending update."""
-    if _inflight_tasks.get(key=scope) is task:
-        _inflight_tasks.pop(key=scope)
-    if task.cancelled():
-        # Cancelled (e.g. bot shutdown): reading result() would raise
-        # CancelledError (a BaseException on 3.11+) out of this callback, and a
-        # pre-shutdown turn is not worth replaying.
-        return
-    try:
-        task.result()
-    except Exception as exc:
-        # Broad on purpose: `result()` re-raises whatever the whole pipeline raised
-        # (LLM, network, file IO), and this runs on the event loop as a done-callback,
-        # so anything escaping here is dropped by asyncio instead of handled.
-        logfire.warn(
-            "Background memory update failed",
-            scope=scope,
-            error_type=type(exc).__name__,
-            _exc_info=exc,
-        )
-    by_subject = _pending_updates.get(key=scope)
-    if not by_subject:
-        _pending_updates.pop(key=scope)
-        return
-    # Oldest source first (dicts keep insertion order), and only one: each replay ends in
-    # this same callback, which picks up the next one.
-    pending = by_subject.pop(next(iter(by_subject)))
-    if not by_subject:
-        _pending_updates.pop(key=scope)
-    if cleared_since(scope=scope, started_at=pending.captured_at):
-        # The durable clear tombstone owns the privacy guarantee. This remains a
-        # best-effort cleanup for store-level clears that only stamped the process.
-        _spawn_db(coro=memory_db.mark_done(scope=scope, token=pending.token))
-        _release_pending_report(pending=pending)
-        return
-    _enqueue_memory_update(
-        scope=scope,
-        subject=pending.subject,
-        transcript=pending.transcript,
-        writer=pending.writer,
-        identity=pending.identity,
-        token=pending.token,
-        report=pending.report,
-    )
-
-
-async def _run_memory_update(  # noqa: PLR0913 -- schedule_memory_update's flavor + payload, and one early exit per way a turn can end
-    scope: str,
-    subject: str,
-    transcript: str,
-    writer: MemoryWriterAI,
-    identity: str,
-    token: int,
-    captured_at: float,
-    report: MemoryWriteReport | None = None,
-) -> None:
+async def _run_memory_update(turn: MemoryTurn) -> None:
     """Reviews the turn's memory notes and, past the raw threshold, consolidates.
 
     The reply.db row is written `pending` at the top (awaited, before the lock) so
@@ -684,7 +331,7 @@ async def _run_memory_update(  # noqa: PLR0913 -- schedule_memory_update's flavo
     when the LLM call itself fails, so the restart sweep retries just that case.
     Consolidation needs no DB row: `raw.md` is its durable, re-entrant queue.
 
-    `captured_at` is when the turn was scheduled, and every clear check runs
+    `turn.captured_at` is when the turn was scheduled, and every clear check runs
     against it rather than a worker-local clock, so a clear that lands while this
     turn is still queued aborts it too.
 
@@ -699,45 +346,42 @@ async def _run_memory_update(  # noqa: PLR0913 -- schedule_memory_update's flavo
     async def settle(summary: MemoryWriteSummary) -> None:
         """Answers the caller's report once, whichever way this turn ends."""
         nonlocal settled
-        if settled or report is None:
+        if settled or turn.report is None:
             return
         settled = True
-        await _report_writes(report=report, summary=summary)
+        await report_writes(report=turn.report, summary=summary)
 
     try:
-        if cleared_since(scope=scope, started_at=captured_at):
+        if cleared_since(scope=turn.scope, started_at=turn.captured_at):
             # Cleared between capture and start: staging the row would hand the
             # restart sweep the very conversation the clear erased, so drop the turn
             # before it writes anything at all.
             return
-        await _safe(
-            coro=_stage_turn(
-                scope=scope,
-                subject=subject,
-                transcript=transcript,
-                identity=identity,
-                token=token,
-                captured_at=captured_at,
+        await safe_db_write(
+            coro=stage_turn(
+                scope=turn.scope,
+                subject=turn.subject,
+                transcript=turn.transcript,
+                identity=turn.identity,
+                token=turn.token,
+                captured_at=turn.captured_at,
             )
         )
-        if cleared_since(scope=scope, started_at=captured_at):
+        if cleared_since(scope=turn.scope, started_at=turn.captured_at):
             # The clear landed while the row was being written. Its durable tombstone
             # owns the DB ordering, so drop the in-memory turn before the review.
             return
-        async with scope_lock(scope=scope), _memory_semaphore():
+        async with scope_lock(scope=turn.scope), memory_semaphore():
             forced = await _review_and_stage(
-                scope=scope,
-                subject=subject,
-                payload=transcript,
-                writer=writer,
-                token=token,
-                captured_at=captured_at,
-                report=settle if report is not None else None,
+                turn=turn, report=settle if turn.report is not None else None
             )
-            if forced is None or not _should_consolidate(scope=scope, forced=forced):
+            if forced is None or not _should_consolidate(scope=turn.scope, forced=forced):
                 return
             await _consolidate_now(
-                scope=scope, started_at=captured_at, writer=writer, identity=identity
+                scope=turn.scope,
+                started_at=turn.captured_at,
+                writer=turn.writer,
+                identity=turn.identity,
             )
     finally:
         # Nothing recorded is the honest answer for every path that reaches here without having
@@ -750,41 +394,36 @@ async def _run_memory_update(  # noqa: PLR0913 -- schedule_memory_update's flavo
         await settle(MemoryWriteSummary())
 
 
-async def _review_and_stage(  # noqa: PLR0913 -- the turn's identity and payload, and one early exit per way a review can end
-    scope: str,
-    subject: str,
-    payload: str,
-    writer: MemoryWriterAI,
-    token: int,
-    captured_at: float,
-    report: MemoryWriteReport | None,
-) -> bool | None:
+async def _review_and_stage(turn: MemoryTurn, report: MemoryWriteReport | None) -> bool | None:
     """Reviews one turn's notes and stages what survives, under the caller's scope lock.
 
     Returns whether consolidation should be FORCED (a forget is waiting), or None when the
     turn is finished and must not consolidate at all. Split out of `_run_memory_update` so the
     lock-holding half reads as one thing; the caller owns only the consolidation decision.
     """
-    transcript, remember_notes, forget_notes = parse_turn_payload(payload=payload)
+    scope = turn.scope
+    transcript, remember_notes, forget_notes = parse_turn_payload(payload=turn.transcript)
     # The subject's source line survives the memory_job round-trip, so a resumed
     # turn stamps the same source; a pre-source row (or the server flavor) parses
     # to None and renders without the source/sharing fields.
-    source = parse_subject_source(subject=subject)
+    source = parse_subject_source(subject=turn.subject)
     # Written before the evaluator runs, and deliberately not undone by its failure: a
     # forget needs no model, and making it wait behind one would let a failed call keep
     # the bot repeating what it was just asked to drop. A retried row therefore writes it
     # twice, which costs nothing: consolidation deletes the fact the first time and finds
     # nothing to delete the second.
     forget_text = render_forget_requests(notes=forget_notes, source=source)
-    if forget_text and not cleared_since(scope=scope, started_at=captured_at):
+    if forget_text and not cleared_since(scope=scope, started_at=turn.captured_at):
         append_raw_entry(scope=scope, entry_text=forget_text)
-    draft = await writer.evaluate(subject=subject, transcript=transcript, notes=remember_notes)
-    if cleared_since(scope=scope, started_at=captured_at):
+    draft = await turn.writer.evaluate(
+        subject=turn.subject, transcript=transcript, notes=remember_notes
+    )
+    if cleared_since(scope=scope, started_at=turn.captured_at):
         # Cleared while this update was in flight; dropping the result beats
         # resurrecting deleted memory. The tombstone already owns the durable
         # ordering; this terminal write is only best-effort cleanup for a
         # process-local store clear.
-        await _safe(coro=memory_db.mark_done(scope=scope, token=token))
+        await safe_db_write(coro=memory_db.mark_done(scope=scope, token=turn.token))
         return None
     if draft is None:
         # The LLM path itself failed: keep the row (payload intact) so the
@@ -795,12 +434,14 @@ async def _review_and_stage(  # noqa: PLR0913 -- the turn's identity and payload
             scope=scope,
             flavor=flavor_of(scope=scope),
         )
-        await _safe(coro=memory_db.mark_failed(scope=scope, token=token, error="evaluate failed"))
+        await safe_db_write(
+            coro=memory_db.mark_failed(scope=scope, token=turn.token, error="evaluate failed")
+        )
         if report is not None and forget_text:
             # The forget is durable regardless of the review, so the reply may say so. Its
             # remembered half is left empty rather than guessed at: the notes that would have
             # filled it are exactly what the failed call was reviewing.
-            await _report_writes(report=report, summary=MemoryWriteSummary(forgotten=forget_notes))
+            await report_writes(report=report, summary=MemoryWriteSummary(forgotten=forget_notes))
         # The forget above is already durable and has nothing to do with the review that
         # failed, so it still gets the immediate pass it was written for. Without this it
         # would wait for an unrelated turn to push the backlog over threshold, and the bot
@@ -827,13 +468,13 @@ async def _review_and_stage(  # noqa: PLR0913 -- the turn's identity and payload
             notes=len(remember_notes),
             candidates=len(draft.observations),
         )
-        await _safe(coro=memory_db.mark_done(scope=scope, token=token))
+        await safe_db_write(coro=memory_db.mark_done(scope=scope, token=turn.token))
         return None
     # The turn is durable in raw.md now; record success before the (best-effort,
     # self-healing) consolidation so a consolidation crash never re-runs the review.
-    await _safe(coro=memory_db.mark_done(scope=scope, token=token))
+    await safe_db_write(coro=memory_db.mark_done(scope=scope, token=turn.token))
     if report is not None:
-        await _report_writes(
+        await report_writes(
             report=report,
             summary=_write_summary(observations=deduped_observations, forgotten=forget_notes),
         )
@@ -885,7 +526,7 @@ async def consolidate_if_needed(scope: str, writer: MemoryWriterAI, identity: st
     digest must never surface), so the caller just spawns it.
     """
     try:
-        async with scope_lock(scope=scope), _memory_semaphore():
+        async with scope_lock(scope=scope), memory_semaphore():
             if not _should_consolidate(scope=scope):
                 return
             _last_consolidation[scope] = time.monotonic()
@@ -967,7 +608,7 @@ async def _consolidate_locked(
         async with asyncio.timeout(MEMORY_CONSOLIDATE_TIMEOUT_SECONDS):
             # Forgets first, so a fact this batch also re-confirms is not deleted right after
             # being written, and so the observation pass sees the tree the deletions left.
-            if not await _apply_forget_buckets(
+            if not await apply_forget_buckets(
                 scope=scope,
                 flavor=flavor,
                 owner=owner,
@@ -994,7 +635,7 @@ async def _consolidate_locked(
                     owner=owner,
                     started_at=started_at,
                     writer=writer,
-                    request_parts=_CompartmentInput(
+                    request_parts=CompartmentInput(
                         raw_entries=buckets.get(compartment, ""),
                         recent_detail=detail_buckets.get(compartment, ""),
                         global_reference=global_reference,
@@ -1018,7 +659,7 @@ async def _consolidate_locked(
         return
     if cleared_since(scope=scope, started_at=started_at):
         return
-    await _update_tone_note(
+    await update_tone_note(
         scope=scope,
         flavor=flavor,
         started_at=started_at,
@@ -1032,7 +673,7 @@ async def _consolidate_locked(
     # free. Synchronous and after the clear guard, like every other write here.
     for compartment in list_compartments(scope=scope):
         sweep_stale_facts(scope=scope, compartment=compartment, today=today_utc())
-    _report_injection_size(scope=scope, flavor=flavor)
+    report_injection_size(scope=scope, flavor=flavor)
     # The consumed batch's content is preserved in the cold-tier detail file; every
     # failure path above returns before this, so it can never retire an unread bucket.
     append_detail(scope=scope, text=raw_entries)
@@ -1042,7 +683,7 @@ async def _consolidate_locked(
     memory_git.enqueue(scope=scope, reason="update")
 
 
-async def _apply_forget_buckets(  # noqa: PLR0913 -- the scope's identity plus the buckets, their stamp and the LLM handle
+async def apply_forget_buckets(  # noqa: PLR0913 -- the scope's identity plus the buckets, their stamp and the LLM handle
     scope: str,
     flavor: MemoryFlavor,
     owner: MemoryOwner,
@@ -1072,7 +713,7 @@ async def _apply_forget_buckets(  # noqa: PLR0913 -- the scope's identity plus t
             started_at=started_at,
             writer=writer,
             deletes_only=True,
-            request_parts=_CompartmentInput(
+            request_parts=CompartmentInput(
                 raw_entries=forget_text, recent_detail="", global_reference="", today=today
             ),
         )
@@ -1081,11 +722,11 @@ async def _apply_forget_buckets(  # noqa: PLR0913 -- the scope's identity plus t
     return True
 
 
-class _CompartmentInput(BaseModel):
+class CompartmentInput(BaseModel):
     """The per-compartment half of a consolidation request, before the store is read.
 
     Split out so a caller can build the parts that differ per compartment without
-    `_compartment_request` growing a dozen positional arguments. The incremental fan-out
+    `compartment_request` growing a dozen positional arguments. The incremental fan-out
     fills all four; a from-scratch rebuild has only the raw bucket and the date.
     """
 
@@ -1104,7 +745,7 @@ async def _consolidate_compartment(  # noqa: PLR0913 -- one compartment's identi
     owner: MemoryOwner,
     started_at: float,
     writer: MemoryWriterAI,
-    request_parts: _CompartmentInput,
+    request_parts: CompartmentInput,
     deletes_only: bool = False,
 ) -> DeltaOutcome | None:
     """Runs and applies one compartment's consolidation; None means the LLM path failed.
@@ -1112,12 +753,12 @@ async def _consolidate_compartment(  # noqa: PLR0913 -- one compartment's identi
     This call only ever sees the evidence routed to the compartment it is writing, which
     is what makes "a guild-locked observation cannot reach `global/`" structural rather
     than a rule the prompt asks the model to follow. The tone note, which is genuinely
-    cross-compartment, is therefore NOT written here — see `_update_tone_note`.
+    cross-compartment, is therefore NOT written here — see `tone.update_tone_note`.
     """
     existing = read_facts(scope=scope, compartment=compartment)
     rendered = render_existing_facts(facts=existing)
     result = await writer.consolidate(
-        request=_compartment_request(
+        request=compartment_request(
             compartment=compartment,
             flavor=flavor,
             existing_facts=rendered,
@@ -1194,54 +835,11 @@ def _record_rejection(scope: str, compartment: str, outcome: DeltaOutcome, store
     )
 
 
-async def _update_tone_note(  # noqa: PLR0913 -- the scope's identity plus the batch, its stamp, and the LLM handle
-    scope: str,
-    flavor: MemoryFlavor,
-    started_at: float,
-    writer: MemoryWriterAI,
-    raw_entries: str,
-    today: str,
-) -> None:
-    """Rewrites the per-user tone note from the WHOLE batch, in its own call.
-
-    Tone is the one tier that is cross-server safe by construction, so it is the one
-    thing that must not be partitioned: nearly half of all observations are
-    `source_only`, and a tone note fed only the `global` bucket would simply stop
-    updating for those conversations.
-
-    That is exactly why it gets its own call rather than riding on the `global`
-    compartment's. A compartment call sees only the evidence routed to the compartment
-    it writes, which is what makes "a guild-locked observation cannot reach `global/`"
-    structural; handing that same call the unpartitioned tone evidence would have
-    demoted the boundary back to a rule the prompt asks the model to follow. Here the
-    deltas are discarded by CODE — this call cannot write a fact anywhere, whatever it
-    returns — so the unpartitioned input is safe by the same structural argument.
-
-    Best-effort throughout: the note is a small always-read tier and the next
-    consolidation repairs a bad write, so a failure never touches the raw batch.
-    """
-    if flavor != "user":
-        return
-    tone_evidence = tone_evidence_from_raw(raw_text=raw_entries)
-    if not tone_evidence:
-        # No tone signal in this batch is the normal case, and an empty output must
-        # never delete the note; only the evidence-complete rebuild may do that.
-        return
-    result = await writer.consolidate(
-        request=_tone_request(
-            existing_tone=read_tone(scope=scope), tone_evidence=tone_evidence, today=today
-        )
-    )
-    if result is None or cleared_since(scope=scope, started_at=started_at):
-        return
-    _write_tone_result(scope=scope, tone_markdown=result.tone_markdown)
-
-
-def _compartment_request(
+def compartment_request(
     compartment: str,
     flavor: MemoryFlavor,
     existing_facts: str,
-    parts: _CompartmentInput,
+    parts: CompartmentInput,
     compact: bool,
 ) -> ConsolidationRequest:
     """Builds one compartment's consolidation request, filling in what never varies.
@@ -1249,7 +847,7 @@ def _compartment_request(
     The compartment note and the flavor's legal sections are what both callers have to get
     right and neither varies with the batch, so they are derived here once; the incremental
     fan-out and the from-scratch rebuild differ only in which blocks they can offer. Neither
-    writes the tone note — `_tone_request` is that call — so `emit_tone` is false for both.
+    writes the tone note — `tone` owns that call — so `emit_tone` is false for both.
     """
     return ConsolidationRequest(
         compartment_note=_compartment_note(compartment=compartment, flavor=flavor),
@@ -1266,27 +864,7 @@ def _compartment_request(
     )
 
 
-def _tone_request(existing_tone: str, tone_evidence: str, today: str) -> ConsolidationRequest:
-    """Builds the tone note's own request, the one consolidation call that writes no fact.
-
-    Its two callers send the same shape and differ only in whether the current note is
-    offered back: the incremental pass merges into it, while the evidence-complete rebuild
-    deliberately ignores it. Sharing the builder is what stops the compartment note — the
-    line telling the model which tier it is writing — drifting between the two.
-    """
-    return ConsolidationRequest(
-        compartment_note="the user's persona-independent tone note, read in every conversation",
-        allowed_sections=(),
-        raw_entries="",
-        existing_tone=existing_tone,
-        tone_evidence=tone_evidence,
-        today=today,
-        compact=False,
-        emit_tone=True,
-    )
-
-
-def _global_first(compartments: set[str]) -> list[str]:
+def global_first(compartments: set[str]) -> list[str]:
     """Orders one run's compartments with `global` leading and the rest by name.
 
     `global` leads because every later compartment is handed its facts as read-only
@@ -1304,7 +882,7 @@ def _compartments_to_run(buckets: dict[str, str]) -> list[str]:
     Only compartments the batch actually routed evidence to: a call with an empty bucket
     has nothing to consolidate.
     """
-    return _global_first(
+    return global_first(
         compartments={
             compartment
             for compartment in buckets
@@ -1329,7 +907,7 @@ def _compartment_note(compartment: str, flavor: MemoryFlavor) -> str:
     return f"memory readable only inside Discord server {compartment.removeprefix('g/')}"
 
 
-def _report_injection_size(scope: str, flavor: MemoryFlavor) -> None:
+def report_injection_size(scope: str, flavor: MemoryFlavor) -> None:
     """Logs when a scope's injectable document approaches or passes the hard cap.
 
     A post-write backstop, not a budget: the read path already stops rendering at the
@@ -1364,372 +942,3 @@ def _report_injection_size(scope: str, flavor: MemoryFlavor) -> None:
             chars=widest,
             cap=MEMORY_INJECTION_MAX_CHARS,
         )
-
-
-def _write_tone_result(scope: str, tone_markdown: str) -> None:
-    """Persists a tone-note call's output when it is acceptable for this scope.
-
-    User scopes only, and only a note starting with the exact `## 語氣偏好` header;
-    an empty or malformed output never deletes the existing note — the tier is
-    best-effort and the next consolidation repairs it. Only `_rebuild_tone_note`,
-    which saw the whole evidence corpus, may clear it.
-    """
-    if flavor_of(scope=scope) != "user":
-        return
-    if not _tone_is_well_formed(tone_markdown=tone_markdown):
-        return
-    write_tone(scope=scope, content=tone_markdown)
-
-
-def _tone_is_well_formed(tone_markdown: str) -> bool:
-    """Whether a tone note carries the exact header the injected tier is contracted to."""
-    return tone_markdown.startswith(_TONE_HEADER)
-
-
-def regeneration_has_evidence(scope: str) -> bool:
-    """Whether any cold-tier evidence exists for a from-scratch rebuild.
-
-    Mirrors the evidence guard inside `regenerate_scope_memory` cheaply (no full
-    window read), so the command can surface "no observations yet" up front
-    instead of scheduling a background rebuild that would silently do nothing.
-    """
-    return bool(read_raw_entries(scope=scope)) or detail_file_bytes(scope=scope) > 0
-
-
-def regeneration_on_cooldown(scope: str) -> bool:
-    """Whether a recent regeneration attempt blocks another one right now."""
-    last_attempt = _last_regeneration.get(scope)
-    if last_attempt is None or cleared_since(scope=scope, started_at=last_attempt):
-        # A clear since the last attempt wiped the memory that cooldown
-        # belonged to; the fresh post-clear state deserves a prompt rebuild.
-        return False
-    return time.monotonic() - last_attempt < MEMORY_REGENERATION_COOLDOWN_SECONDS
-
-
-def schedule_memory_regeneration(scope: str, writer: MemoryWriterAI, identity: str) -> bool:
-    """Starts a background rebuild of the scope's memory without blocking the command.
-
-    Returns False when a rebuild is already in flight for this scope (so the
-    caller can report "still rebuilding" instead of double-scheduling the
-    rebuild); True when a fresh background task was started.
-    """
-    running = _regeneration_tasks.get(key=scope)
-    if running is not None and not running.done():
-        return False
-    task = asyncio.create_task(
-        regenerate_scope_memory(scope=scope, writer=writer, identity=identity)
-    )
-    _regeneration_tasks.set(key=scope, value=task)
-    task.add_done_callback(
-        lambda finished: _finish_memory_regeneration(scope=scope, task=finished)
-    )
-    return True
-
-
-def _finish_memory_regeneration(scope: str, task: asyncio.Task[RegenerationReport]) -> None:
-    """Clears the in-flight slot and logs failures of a background rebuild."""
-    if _regeneration_tasks.get(key=scope) is task:
-        _regeneration_tasks.pop(key=scope)
-    if task.cancelled():
-        # Cancelled (e.g. bot shutdown): reading result() would raise
-        # CancelledError out of this callback, and an aborted rebuild leaves the
-        # existing memory untouched, so there is nothing to recover.
-        return
-    try:
-        task.result()
-    except Exception as exc:
-        # Broad on purpose: this is a done-callback boundary, so anything the
-        # rebuild raised must be swallowed here or asyncio drops it silently.
-        logfire.error(
-            "Background memory regeneration crashed",
-            scope=scope,
-            error_type=type(exc).__name__,
-            _exc_info=exc,
-        )
-
-
-async def regenerate_scope_memory(
-    scope: str, writer: MemoryWriterAI, identity: str
-) -> RegenerationReport:
-    """Rebuilds every compartment from cold-tier evidence alone.
-
-    The existing facts are deliberately NOT fed to the model: the rebuild distills the
-    detail tail window plus any unconsumed raw entries from scratch, e.g. to redo an
-    unsatisfying consolidation with another model. Facts the rebuild did not re-emit are
-    then deleted, which is the one path allowed to lose most of a compartment at once:
-    replacing the whole set is what it is for, and that exemption is also what let the
-    #408 compartment migration run through here instead of needing a writer of its own.
-    `scripts/regen_memories.py` drives the same path offline.
-
-    On an LLM failure the compartment is left exactly as it was, and the raw batch is
-    retired only when every compartment rebuilt. The report carries what the run removed
-    unread whichever way it ended, so a rebuild that gave up on its third compartment
-    still accounts for what the first two destroyed.
-    """
-    started_at = time.monotonic()
-    unreadable_removed = 0
-    async with scope_lock(scope=scope), _memory_semaphore():
-        if regeneration_on_cooldown(scope=scope):
-            # Invocations queued behind a held lock all pass the command-level
-            # cooldown check before the first one stamps the attempt; the
-            # re-check under the lock keeps the per-scope limit on the rewrite.
-            return RegenerationReport(result="cooldown")
-        flavor = flavor_of(scope=scope)
-        owner = parse_identity(identity=identity, fallback_owner_id=scope_owner_id(scope=scope))
-        raw_entries = read_raw_entries(scope=scope)
-        recent_detail = read_detail_tail(scope=scope, max_chars=MEMORY_DETAIL_CONTEXT_MAX_CHARS)
-        # Detail entries are retired raw entries verbatim with the same
-        # `## <ISO timestamp>` headers, so the combined corpus (oldest first)
-        # slots into the raw-entries consolidation input unchanged.
-        evidence = "\n\n".join(part for part in (recent_detail, raw_entries) if part)
-        if not evidence:
-            return RegenerationReport(result="no_evidence")
-        # Recorded at attempt time, not success time, so repeated LLM failures
-        # are rate-limited by the same cooldown.
-        _last_regeneration[scope] = time.monotonic()
-        buckets = partition_raw_entries(raw_text=evidence, flavor=flavor)
-        today = datetime.now(UTC).date().isoformat()
-        # Every compartment that has evidence, plus every one that still has files, so a
-        # compartment whose evidence is gone is emptied rather than left stale.
-        compartments = _compartments_to_rebuild(scope=scope, buckets=buckets)
-        try:
-            # Bounded as a whole like the incremental fan-out, and for the same reason: the
-            # individual calls carry no deadline of their own (`constants.py` has why), so
-            # this is the only thing standing between a stuck rebuild and a scope lock held
-            # for as long as the client will keep one compartment's request alive.
-            async with asyncio.timeout(MEMORY_CONSOLIDATE_TIMEOUT_SECONDS):
-                for compartment in compartments:
-                    raw_bucket = buckets.get(compartment, "")
-                    if not raw_bucket and not read_facts(scope=scope, compartment=compartment):
-                        # A leftover directory with nothing to distil and nothing to keep:
-                        # the model would be handed an empty corpus and could only answer
-                        # with an empty batch, so the prune alone reaches the same state.
-                        # It also removes the emptied directory, which is what stops the
-                        # leftover costing another call — and another way to fail the
-                        # compartments that do have something — on every later rebuild.
-                        unreadable_removed += _prune_rebuilt_compartment(
-                            scope=scope, compartment=compartment, keep=set()
-                        )
-                        continue
-                    result = await writer.consolidate(
-                        request=_compartment_request(
-                            compartment=compartment,
-                            flavor=flavor,
-                            existing_facts="",
-                            parts=_CompartmentInput(
-                                raw_entries=raw_bucket,
-                                recent_detail="",
-                                global_reference="",
-                                today=today,
-                            ),
-                            compact=True,
-                        )
-                    )
-                    if result is None:
-                        # The LLM path logs the cause but not the scope, and the command
-                        # already told the user a rebuild was scheduled, so this is its
-                        # only attribution.
-                        logfire.warn(
-                            "Memory regeneration LLM call failed; memory left untouched",
-                            scope=scope,
-                            compartment=compartment,
-                        )
-                        return RegenerationReport(
-                            result="failed", unreadable_removed=unreadable_removed
-                        )
-                    if cleared_since(scope=scope, started_at=started_at):
-                        return RegenerationReport(
-                            result="failed", unreadable_removed=unreadable_removed
-                        )
-                    unreadable_removed += _replace_compartment(
-                        scope=scope,
-                        compartment=compartment,
-                        flavor=flavor,
-                        owner=owner,
-                        result=result,
-                    )
-                await _reapply_forgets(
-                    scope=scope,
-                    flavor=flavor,
-                    owner=owner,
-                    started_at=started_at,
-                    writer=writer,
-                    evidence=evidence,
-                    today=today,
-                )
-                await _rebuild_tone_note(
-                    scope=scope,
-                    flavor=flavor,
-                    started_at=started_at,
-                    writer=writer,
-                    evidence=evidence,
-                    today=today,
-                )
-        except TimeoutError:
-            logfire.warn(
-                "Memory regeneration timed out", scope=scope, compartments=len(compartments)
-            )
-            return RegenerationReport(result="failed", unreadable_removed=unreadable_removed)
-        _report_injection_size(scope=scope, flavor=flavor)
-        if raw_entries:
-            # The rebuild consumed the raw batch; retire it to the cold tier
-            # exactly like a consolidation so it cannot be re-ingested.
-            append_detail(scope=scope, text=raw_entries)
-            clear_raw(scope=scope)
-        memory_git.enqueue(scope=scope, reason="rebuild")
-        return RegenerationReport(result="regenerated", unreadable_removed=unreadable_removed)
-
-
-async def _reapply_forgets(  # noqa: PLR0913 -- the scope's identity plus the corpus, its stamp, and the LLM handle
-    scope: str,
-    flavor: MemoryFlavor,
-    owner: MemoryOwner,
-    started_at: float,
-    writer: MemoryWriterAI,
-    evidence: str,
-    today: str,
-) -> None:
-    """Re-runs every forget request in the corpus against the freshly rebuilt facts.
-
-    A rebuild derives facts from evidence rather than from the current facts, and the
-    observation a forget was aimed at is still sitting in `detail.md` verbatim: consolidation
-    retires the raw batch there and never prunes it. So the rebuild re-creates exactly what
-    the user asked to have removed, and `/memory regenerate` quietly undoes every forget they
-    ever asked for.
-
-    Replaying the requests afterwards fixes that without weakening anything: each runs as its
-    own `deletes_only` call, the same shape the incremental path uses, so the forget's own
-    sentence still cannot be written anywhere. Feeding the requests INTO the rebuild instead
-    would have put a possibly-private sentence in front of a call whose whole job is creating
-    facts, which is the one thing `deletes_only` exists to prevent.
-
-    Best-effort: the rebuild has already landed by this point, and a failure here leaves a
-    resurrected fact rather than a broken store. The next forget removes it again.
-    """
-    await _apply_forget_buckets(
-        scope=scope,
-        flavor=flavor,
-        owner=owner,
-        started_at=started_at,
-        writer=writer,
-        buckets=partition_forget_requests(
-            raw_text=evidence, compartments=tuple(list_compartments(scope=scope))
-        ),
-        today=today,
-    )
-
-
-async def _rebuild_tone_note(  # noqa: PLR0913 -- the scope's identity plus the corpus, its stamp, and the LLM handle
-    scope: str,
-    flavor: MemoryFlavor,
-    started_at: float,
-    writer: MemoryWriterAI,
-    evidence: str,
-    today: str,
-) -> None:
-    """Rebuilds the tone note from the whole evidence corpus, in its own call.
-
-    Unlike an incremental consolidation — whose empty tone output only means "no tone
-    signal in this batch" — this pass saw everything, so no signal anywhere means a
-    surviving note is stale and would keep injecting a preference the evidence no longer
-    supports. This is the only path allowed to delete the note.
-    """
-    if flavor != "user":
-        return
-    tone_evidence = tone_evidence_from_raw(raw_text=evidence)
-    result = (
-        None
-        if not tone_evidence
-        else await writer.consolidate(
-            # No `existing_tone`: this pass saw the whole corpus, so it rewrites the note
-            # from the evidence rather than merging into what is already there.
-            request=_tone_request(existing_tone="", tone_evidence=tone_evidence, today=today)
-        )
-    )
-    if cleared_since(scope=scope, started_at=started_at):
-        return
-    if result is None or not result.tone_markdown:
-        clear_tone(scope=scope)
-        return
-    _write_tone_result(scope=scope, tone_markdown=result.tone_markdown)
-
-
-def _compartments_to_rebuild(scope: str, buckets: dict[str, str]) -> list[str]:
-    """Returns every compartment a rebuild touches, `global` first.
-
-    Compartments that still hold files but have no surviving evidence are included so
-    the rebuild empties them; leaving them alone would keep pre-rebuild facts visible
-    alongside the new ones with no evidence behind them. Touching one does not always
-    mean consolidating it: an entry that turns out to hold neither evidence nor a
-    readable fact is pruned without a model call (`regenerate_scope_memory` has the why).
-    """
-    return _global_first(
-        compartments={GLOBAL_COMPARTMENT, *buckets, *list_compartments(scope=scope)}
-    )
-
-
-def _replace_compartment(
-    scope: str,
-    compartment: str,
-    flavor: MemoryFlavor,
-    owner: MemoryOwner,
-    result: ConsolidatedMemory,
-) -> int:
-    """Replaces a compartment's contents with a from-scratch rebuild's facts.
-
-    Applies the batch first and prunes afterwards, so a fact the rebuild kept is never
-    momentarily absent. What survives is decided by the ids the batch actually WROTE, not
-    by what is left on disk: a rebuild says "this fact is gone" by simply not mentioning
-    it, so comparing against the post-apply state would only re-delete what the batch
-    already deleted and leave every stale fact standing.
-
-    The mass-delete guard is off here for the reason it was skipped by the old whole-file
-    rebuild: replacing the entire set is what this path is for.
-    """
-    outcome = apply_deltas(
-        scope=scope,
-        compartment=compartment,
-        flavor=flavor,
-        deltas=result.deltas,
-        owner=owner,
-        allow_mass_delete=True,
-    )
-    return _prune_rebuilt_compartment(
-        scope=scope, compartment=compartment, keep=set(outcome.written)
-    )
-
-
-def _prune_rebuilt_compartment(scope: str, compartment: str, keep: set[str]) -> int:
-    """Reduces a rebuilt compartment to `keep`, reporting what it left and what it took.
-
-    The prune reads the directory rather than the facts read back from it, so a file no
-    reader can parse cannot outlive a rebuild that reports the compartment replaced
-    (`prune_compartment` carries the why). What it could not account for is reported
-    here instead, since the store never removes a file it did not write.
-
-    What it DID remove unread is reported here too, and returned for the offline
-    rebuild's own report. Renaming a section or a durability value makes every fact
-    carrying the old one unparsable, so the next rebuild of a scope drops all of them in
-    one pass; a run that says only what it spared reads as one that destroyed nothing.
-
-    Shared with the skip path, which prunes a compartment it never handed to the model,
-    so a file the store never wrote is named there on the same terms — and so is the
-    unreadable one, which is the ONLY thing that path ever removes: a compartment reaches
-    it precisely when nothing in it could be read.
-    """
-    pruned = prune_compartment(scope=scope, compartment=compartment, keep=keep)
-    if pruned.unaccounted:
-        logfire.warn(
-            "Memory rebuild left files it cannot account for",
-            scope=scope,
-            compartment=compartment,
-            files=pruned.unaccounted,
-        )
-    if pruned.unreadable:
-        logfire.warn(
-            "Memory rebuild removed fact files it could not read",
-            scope=scope,
-            compartment=compartment,
-            files=pruned.unreadable,
-        )
-    return len(pruned.unreadable)
